Move type utils from types.ts to their own file
