+++ packages/core/src/index.ts
@@ -7,3 +7,4 @@ export * from './render';
 export * from './defaultConfig';
 export * from './convertConfig';
 export * from './typecheck';
+export * from './typeutils';
